remove pkglink and use pkl instead
